Improve semantics, refactor CSS, and change svg to png
